median changes for size_t

+++ src/median.h
@@ -3,11 +3,11 @@
 
 
 
-void ColMedian(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes, double *resultsSE);
-void ColMedian_noSE(double *data, int rows, int cols, int *cur_rows, double *results, int nprobes);
+void ColMedian(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes, double *resultsSE);
+void ColMedian_noSE(double *data, size_t rows, size_t cols, int *cur_rows, double *results, size_t nprobes);
 
-void colmedian(double *data, int rows, int cols, double *results, double *resultsSE);
-void colmedian_no_copy(double *data, int rows, int cols, double *results, double *resultsSE);
+void colmedian(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
+void colmedian_no_copy(double *data, size_t rows, size_t cols, double *results, double *resultsSE);
